@@ -1,70 +1,59 @@
 {
-  "name": "cv-ressad-nour-official",
+  "name": "cv.ressad.nour",
   "version": "1.0.0",
-  "private": true,
+  "license": "MIT",
   "scripts": {
     "ng": "ng",
-    "build": "npm run env -s && ng build --prod",
-    "start": "ng serve --aot --proxy-config proxy.conf.js",
-    "lint": "ng lint --type-check && stylelint \"src/**/*.scss\" --syntax scss && htmlhint \"src\" --config .htmlhintrc",
+    "start": "ng serve",
+    "build": "ng build --prod",
     "test": "ng test",
-    "test:ci": "npm run lint -s && ng test --single-run --code-coverage",
-    "e2e": "ng e2e",
-    "translations:extract": "ngx-translate-extract --input ./src --output ./src/translations/template.json --format=json --clean -sort --marker extract",
-    "docs": "hads ./docs -o",
-    "env": "ngx-scripts env2json npm_package_version",
-    "generate": "ng generate"
+    "lint": "ng lint",
+    "e2e": "ng e2e"
   },
+  "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
-    "@ngx-translate/core": "^9.0.1",
-    "@angular/service-worker": "^1.0.0-beta.16",
-    "@ng-bootstrap/ng-bootstrap": "^1.0.0-beta.5",
-    "bootstrap": "4.0.0-beta.2",
-    "font-awesome": "^4.7.0",
+    "@angular/animations": "^5.2.0",
+    "@angular/common": "^5.2.0",
+    "@angular/compiler": "^5.2.0",
+    "@angular/core": "^5.2.0",
+    "@angular/forms": "^5.2.0",
+    "@angular/http": "^5.2.0",
+    "@angular/platform-browser": "^5.2.0",
+    "@angular/platform-browser-dynamic": "^5.2.0",
+    "@angular/router": "^5.2.0",
+    "@ng-bootstrap/ng-bootstrap": "^1.0.0-beta.9",
+    "@ngx-translate/core": "^9.1.1",
+    "@ngx-translate/http-loader": "^2.0.1",
+    "@nicky-lenaers/ngx-scroll-to": "^0.5.0",
+    "bootstrap": "^4.0.0",
     "core-js": "^2.4.1",
-    "lodash": "^4.17.4",
-    "rxjs": "5.5.2",
-    "zone.js": "^0.8.14"
+    "font-awesome": "^4.7.0",
+    "jquery": "^1.9.1",
+    "ng2-animate-on-scroll": "^1.0.8",
+    "ngx-bootstrap": "^2.0.1",
+    "popper.js": "^1.12.9",
+    "rxjs": "^5.5.6",
+    "zone.js": "^0.8.19"
   },
   "devDependencies": {
-    "@angular/cli": "~1.5.0",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^5.0.0",
-    "@biesbjerg/ngx-translate-extract": "^2.3.1",
-    "@ngx-rocket/scripts": "^1.1.0",
-    "@types/jasmine": "^2.5.52",
-    "@types/jasminewd2": "^2.0.2",
-    "@types/lodash": "^4.14.52",
-    "@types/node": "^8.0.10",
-    "codelyzer": "~4.0.1",
-    "hads": "^1.3.4",
-    "htmlhint": "^0.9.13",
-    "https-proxy-agent": "^2.0.0",
+    "@angular/cli": "1.6.5",
+    "@angular/compiler-cli": "^5.2.0",
+    "@angular/language-service": "^5.2.0",
+    "@types/jasmine": "~2.8.3",
+    "@types/jasminewd2": "~2.0.2",
+    "@types/node": "~6.0.60",
+    "codelyzer": "^4.0.1",
     "jasmine-core": "~2.8.0",
-    "jasmine-spec-reporter": "^4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "^2.2.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~2.0.0",
+    "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "^1.1.0",
-    "karma-junit-reporter": "^1.2.0",
-    "protractor": "~5.2.0",
-    "puppeteer": "^0.13.0",
-    "stylelint": "~8.3.1",
-    "stylelint-config-recommended-scss": "~2.0.0",
-    "stylelint-config-standard": "~18.0.0",
-    "stylelint-scss": "~2.1.0",
-    "ts-node": "~3.3.0",
-    "tslint": "~5.8.0",
-    "typescript": "~2.4.2"
+    "karma-jasmine": "~1.1.0",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "protractor": "~5.1.2",
+    "ts-node": "~4.1.0",
+    "tslint": "~5.9.1",
+    "typescript": "~2.5.3"
   }
 }
