@@ -1,14 +1,14 @@
 import { AppPage } from './app.po';
 
-describe('app', () => {
+describe('cv.ressad.nour App', () => {
   let page: AppPage;
 
   beforeEach(() => {
     page = new AppPage();
   });
 
-  it('should display hello message', () => {
+  it('should display welcome message', () => {
     page.navigateTo();
-    expect(page.getParagraphText()).toEqual('Hello world !');
+    expect(page.getParagraphText()).toEqual('Welcome to app!');
   });
 });
