@@ -1,7 +1,5 @@
 import { Component, OnInit } from '@angular/core';
 
-import { environment } from '../../environments/environment';
-
 @Component({
   selector: 'app-about',
   templateUrl: './about.component.html',
@@ -9,10 +7,9 @@ import { environment } from '../../environments/environment';
 })
 export class AboutComponent implements OnInit {
 
-  version: string = environment.version;
-
   constructor() { }
 
-  ngOnInit() { }
+  ngOnInit() {
+  }
 
 }
