@@ -1,5 +1,5 @@
-// `.env.json` is generated by the `npm run build` command
-import * as env from './.env.json';
+// `.env.ts` is generated by the `npm run env` command
+import env from './.env';
 
 export const environment = {
   production: true,
