@@ -1,11 +1,7 @@
-/*
- * Extra typings definitions
- */
-
 // Allow .json files imports
 declare module '*.json';
 
-// SystemJS module definition
+/* SystemJS module definition */
 declare var module: NodeModule;
 interface NodeModule {
   id: string;
